next.js Fortunes talking to Postgres
