+++ src/BenchmarksApps/TechEmpower/nextjs/next.config.js
@@ -1,4 +1,5 @@
 /** @type {import('next').NextConfig} */
 module.exports = {
   output: 'standalone',
+  reactStrictMode: true
 }
