+++ src/BenchmarksApps/TechEmpower/nextjs/package.json
@@ -15,8 +15,12 @@
     "eslint": "8.42.0",
     "eslint-config-next": "13.4.0",
     "next": "13.4.0",
+    "pg": "^8.11.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "typescript": "5.1.3"
+  },
+  "devDependencies": {
+    "@types/pg": "^8.10.2"
   }
 }
